Inline Case files for SuppressingSpec
