+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/SuppressingSpec.kt
@@ -1,7 +1,6 @@
 package io.gitlab.arturbosch.detekt.rules
 
-import io.github.detekt.test.utils.compileForTest
-import io.github.detekt.test.utils.resourceAsPath
+import io.github.detekt.test.utils.compileContentForTest
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.core.rules.visitFile
 import io.gitlab.arturbosch.detekt.rules.complexity.ComplexCondition
@@ -18,7 +17,57 @@ class SuppressingSpec {
 
     @Test
     fun `all findings are suppressed on element levels`() {
-        val ktFile = compileForTest(resourceAsPath("SuppressedByElementAnnotation.kt"))
+        @Suppress("KotlinConstantConditions")
+        val code = """
+            @SuppressWarnings("LongParameterList")
+            fun lpl(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int) = (a + b + c + d + e + f).apply {
+                assert(false) { "FAILED TEST" }
+            }
+            
+            @SuppressWarnings("ComplexCondition")
+            class SuppressedElements {
+            
+                @SuppressWarnings("LongParameterList")
+                fun lpl(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int) = (a + b + c + d + e + f).apply {
+                    assert(false) { "FAILED TEST" }
+                }
+            
+                @SuppressWarnings("ComplexCondition")
+                fun cc() {
+                    if (this is SuppressedElements && this !is Any && this is Nothing && this is SuppressedElements) {
+                        assert(false) { "FAIL" }
+                    }
+                }
+            
+                @SuppressWarnings("LongMethod")
+                fun lm() {
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    assert(false) { "FAILED TEST" }
+                }
+            
+            }
+        """.trimIndent()
+        val ktFile = compileContentForTest(code)
         val ruleSet = RuleSet("Test", listOf(LongMethod(), LongParameterList(), ComplexCondition()))
 
         val findings = ruleSet.visitFile(ktFile)
@@ -28,7 +77,53 @@ class SuppressingSpec {
 
     @Test
     fun `all findings are suppressed on file levels`() {
-        val ktFile = compileForTest(resourceAsPath("SuppressedElementsByFileAnnotation.kt"))
+        @Suppress("KotlinConstantConditions")
+        val code = """
+            @file:Suppress("LongMethod", "LongParameterList", "ComplexCondition")
+            
+            fun lpl2(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int) = (a + b + c + d + e + f).apply {
+                assert(false) { "FAILED TEST" }
+            }
+            
+            class SuppressedElements2 {
+            
+                fun lpl(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int) = (a + b + c + d + e + f).apply {
+                    assert(false) { "FAILED TEST" }
+                }
+            
+                fun cc() {
+                    if (this is SuppressedElements2 && this !is Any && this is Nothing && this is SuppressedElements2) {
+                        assert(false) { "FAIL" }
+                    }
+                }
+            
+                fun lm() {
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    assert(false) { "FAILED TEST" }
+                }
+            }
+        """.trimIndent()
+        val ktFile = compileContentForTest(code)
         val ruleSet = RuleSet("Test", listOf(LongMethod(), LongParameterList(), ComplexCondition()))
 
         val findings = ruleSet.visitFile(ktFile)
@@ -38,7 +133,49 @@ class SuppressingSpec {
 
     @Test
     fun `all findings are suppressed on class levels`() {
-        val ktFile = compileForTest(resourceAsPath("SuppressedElementsByClassAnnotation.kt"))
+        @Suppress("KotlinConstantConditions")
+        val code = """
+            @Suppress("LongMethod", "LongParameterList", "ComplexCondition")
+            class SuppressedElements3 {
+            
+                fun lpl(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int) = (a + b + c + d + e + f).apply {
+                    assert(false) { "FAILED TEST" }
+                }
+            
+                fun cc() {
+                    if (this is SuppressedElements3 && this !is Any && this is Nothing && this is SuppressedElements3) {
+                        assert(false) { "FAIL" }
+                    }
+                }
+            
+                fun lm() {
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    lpl(1, 2, 3, 4, 5, 6)
+                    assert(false) { "FAILED TEST" }
+                }
+            }
+        """.trimIndent()
+
+        val ktFile = compileContentForTest(code)
         val ruleSet = RuleSet("Test", listOf(LongMethod(), LongParameterList(), ComplexCondition()))
 
         val findings = ruleSet.visitFile(ktFile)
@@ -57,8 +194,17 @@ class SuppressingSpec {
 
     @Test
     fun `should suppress StringLiteralDuplication on class level`() {
-        val path = resourceAsPath("SuppressStringLiteralDuplication.kt")
+        @Suppress("UnusedEquals")
+        val code = """
+            @Suppress("StringLiteralDuplication")
+            class Duplication {
+                var s1 = "lorem"
+                fun f(s: String = "lorem") {
+                    s1 == "lorem"
+                }
+            }
+        """.trimIndent()
 
-        assertThat(StringLiteralDuplication().lint(path)).isEmpty()
+        assertThat(StringLiteralDuplication().lint(code)).isEmpty()
     }
 }
